+++ packages/patrol/lib/src/native/native_automator2.dart
@@ -652,6 +652,7 @@ class NativeAutomator2 {
     required Offset to,
     int steps = 12,
     String? appId,
+    bool enablePatrolLog = true,
   }) async {
     assert(from.dx >= 0 && from.dx <= 1);
     assert(from.dy >= 0 && from.dy <= 1);
@@ -660,6 +661,7 @@ class NativeAutomator2 {
 
     await _wrapRequest(
       'swipe',
+      enablePatrolLog: enablePatrolLog,
       () => _client.swipe(
         SwipeRequest(
           startX: from.dx,
@@ -693,9 +695,17 @@ class NativeAutomator2 {
   /// await tester.swipeBack(dy: 0.8); // Swipe back at 1/5 height of the screen
   /// await tester.swipeBack(); // Swipe back at the center of the screen
   /// ```
-  Future<void> swipeBack({double dy = 0.5, String? appId}) {
+  Future<void> swipeBack({double dy = 0.5, String? appId}) async {
     assert(dy >= 0.0 && dy <= 1.0, 'dy must be between 0.0 and 1.0');
-    return swipe(from: Offset(0, dy), to: Offset(1, dy), appId: appId);
+    await _wrapRequest(
+      'swipeBack',
+      () => swipe(
+        from: Offset(0, dy),
+        to: Offset(1, dy),
+        appId: appId,
+        enablePatrolLog: false,
+      ),
+    );
   }
 
   /// Simulates pull-to-refresh gesture.
@@ -717,16 +727,20 @@ class NativeAutomator2 {
     Offset from = const Offset(0.5, 0.5),
     Offset to = const Offset(0.5, 0.9),
     int steps = 50,
-  }) {
+  }) async {
     assert(from.dx >= 0 && from.dx <= 1);
     assert(from.dy >= 0 && from.dy <= 1);
     assert(to.dx >= 0 && to.dx <= 1);
     assert(to.dy >= 0 && to.dy <= 1);
 
-    return swipe(
-      from: Offset(from.dx, from.dy),
-      to: Offset(to.dx, to.dy),
-      steps: steps,
+    await _wrapRequest(
+      'pullToRefresh',
+      () => swipe(
+        from: Offset(from.dx, from.dy),
+        to: Offset(to.dx, to.dy),
+        steps: steps,
+        enablePatrolLog: false,
+      ),
     );
   }