+++ packages/patrol_cli/test/src/mocks.dart
@@ -5,8 +5,13 @@ import 'package:mocktail/mocktail.dart';
 import 'package:patrol_cli/src/analytics/analytics.dart';
 import 'package:patrol_cli/src/android/android_test_backend.dart';
 import 'package:patrol_cli/src/base/logger.dart' as logger;
+import 'package:patrol_cli/src/compatibility_checker/compatibility_checker.dart';
+import 'package:patrol_cli/src/dart_defines_reader.dart';
 import 'package:patrol_cli/src/devices.dart';
 import 'package:patrol_cli/src/ios/ios_test_backend.dart';
+import 'package:patrol_cli/src/pubspec_reader.dart';
+import 'package:patrol_cli/src/test_bundler.dart';
+import 'package:patrol_cli/src/test_finder.dart';
 import 'package:process/process.dart' as process;
 import 'package:pub_updater/pub_updater.dart' as pub;
 
@@ -29,6 +34,16 @@ class MockDeviceFinder extends Mock implements DeviceFinder {}
 class MockAndroidTestBackend extends Mock implements AndroidTestBackend {}
 
 class MockIOSTestBackend extends Mock implements IOSTestBackend {}
+
+class MockTestFinder extends Mock implements TestFinder {}
+
+class MockTestBundler extends Mock implements TestBundler {}
+
+class MockDartDefinesReader extends Mock implements DartDefinesReader {}
+
+class MockPubspecReader extends Mock implements PubspecReader {}
+
+class MockCompatibilityChecker extends Mock implements CompatibilityChecker {}
 
 class MockAnalytics extends Mock implements Analytics {
   @override