+++ packages/patrol_cli/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Add support for de, fr and pl locales for native methods that operates on strings (#2659)
 - Add support for gallery permission dialog on iOS 17 (#2659)
+- Add printing paths to the APKs after `patrol build` command. (#2685)
 - Add support for `--no-tree-shake-icons` flag in cli commands (#2704)
 
 ## 3.9.0